third_party/blink/web_tests/platform/generic/fast/events/resize-events-count-expected.txt

Test how many resize events are emitted during window resize.
PASS successfullyParsed is true

TEST COMPLETE

PASS resizeEventCount is 1
